{{ cookiecutter.project_name }}/{{ cookiecutter.project_name }}/celery_app.py
from __future__ import absolute_import

import os

from celery import Celery
from django.conf import settings

# set the default Django settings module for the 'celery' program.
os.environ.setdefault('DJANGO_SETTINGS_MODULE',
                      '{{ cookiecutter.project_name }}.settings')

app = Celery('{{ cookiecutter.project_name }}')
app.config_from_object('django.conf:settings')
app.autodiscover_tasks(lambda: settings.INSTALLED_APPS)
